test/inspector/debugger/session-stop-expected.txt

Checks V8InspectorSession::stop

Running test: testSessionStopResumesPause
Evaluation returned: 42

Running test: testSessionStopResumesInstrumentationPause
Paused: instrumentation
Evaluation returned: 42

Running test: testSessionStopDisablesDebugger
Pause error(?): Debugger agent is not enabled

Running test: testSessionStopDisallowsReenabling
Pause error(?) after stop: Debugger agent is not enabled
Pause error(?) after re-enable: Debugger agent is not enabled
